fix(telemetry): satisfy composition ratchets
